backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
distro==1.9.0
fastapi==0.116.1
Flask==3.1.3
flask-cors==6.0.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==0.3.27
langchain-core==0.3.83
langchain-text-splitters==0.3.11
langsmith==0.7.22
MarkupSafe==3.0.3
openai==2.30.0
orjson==3.11.7
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.0
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.47.3
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.35.0
Werkzeug==3.1.7
xxhash==3.6.0
zstandard==0.25.0
